feat : 지난주 받을 로또 번호을 보너스 번호까지 추가합니다.
